+++ log4j2.properties
@@ -33,15 +33,15 @@ rootLogger.appenderRef.rolling.ref = FILE
 #logger.block.level = trace
 
 ## see more details for RNS, ReticulumPeer
-#logger.RNS.name = org.qortal.network.RNS
+#logger.RNS.name = org.qortal.network.reticulum.RNS
 #logger.RNS.level = debug
-#logger.ReticulumPeer.name = org.qortal.network.ReticulumPeer
+#logger.ReticulumPeer.name = org.qortal.network.reticulum.ReticulumPeer
 #logger.ReticulumPeer.level = debug
 
 ## add DEBUG for RNS components
-#logger.rns.name = org.qortal.network.RNS
+#logger.rns.name = org.qortal.network.reticulum.RNS
 #logger.rns.level = DEBUG
-#logger.reticulumpeer.name = org.qortal.network.ReticulumPeer
+#logger.reticulumpeer.name = org.qortal.network.reticulum.ReticulumPeer
 #logger.reticulumpeer.level = DEBUG
 
 ### end additional debug options
