+++ src/main/java/org/qortal/api/resource/PeersResource.java
@@ -28,8 +28,8 @@ import org.qortal.network.Peer;
 import org.qortal.network.PeerFactory;
 import org.qortal.network.PeerAddress;
 import org.qortal.network.PeerAddressFactory;
-import org.qortal.network.RNS;
-import org.qortal.network.ReticulumPeer;
+import org.qortal.network.reticulum.RNS;
+import org.qortal.network.reticulum.ReticulumPeer;
 import org.qortal.repository.DataException;
 import org.qortal.repository.Repository;
 import org.qortal.repository.RepositoryManager;
